move datatags to external file
