+++ DataTags/datatags.js
@@ -1,0 +1,10 @@
+// Copyright (c) 2020 Mitchell Adair
+// 
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+
+const twitch = require('./Twitch/twitchDataTags');
+
+module.exports = [
+    ...twitch,
+]
